requirements_rag.txt
# Legal RAG System Requirements

# Core dependencies
psycopg2-binary>=2.9.9
pgvector>=0.2.0

# PDF Processing
pdfplumber>=0.9.0
pypdf>=3.0.0
PyPDF2>=3.0.0

# NLP & ML
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
spacy>=3.7.0
nltk>=3.8.1
rank-bm25>=0.2.2

# LLM & LangChain
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.6.0
anthropic>=0.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.1.0

# API (optional)
fastapi>=0.109.0
uvicorn>=0.27.0

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.1

# Development
pytest>=7.4.0
black>=23.12.0
flake8>=6.1.0

# Database
sqlalchemy>=2.0.23

# Additional utilities
tqdm>=4.66.0
python-dateutil>=2.8.2
regex>=2023.12.25
